requirements.txt
asgiref==3.4.1
coverage==6.2
Django==3.2.10
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==10.0.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==4.0.1
